+++ app/(route)/sellAndBuy/page.tsx
@@ -1,39 +1,15 @@
 'use client';
 
-import React, { useState, ChangeEvent } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
 import Button from '@/components/common/Button';
 import theme from '@/styles/theme';
-
-const StyledButton = styled(Button)`
-    color: ${theme.colors.text.primary}; // 텍스트 색상을 검정으로 설정합니다.
-`;
+import BidSectionPage from './BidSection';
+import BuySectionPage from './BuySection';
 
 const SellAndBuyPage = () => {
+
     const [view, setView] = useState('buy');
-    const [bidPrice, setBidPrice] = useState('');
-    const [bidError, setBidError] = useState('');
-    const [expiry, setExpiry] = useState<number | null>(null);
-    const [expiryDate, setExpiryDate] = useState('');
-
-    const handleBidPriceChange = (e: ChangeEvent<HTMLInputElement>) => {
-        const value = e.target.value;
-        if (/^\d+$/.test(value) && parseInt(value) >= 20000 && parseInt(value) % 1000 === 0) {
-            setBidPrice(value);
-            setBidError('');
-        } else {
-            setBidPrice(value);
-            setBidError('2만원부터 천원단위로 입력하세요.');
-        }
-    };
-
-    const handleExpiryChange = (days: number) => {
-        const today = new Date();
-        today.setDate(today.getDate() + days);
-        const formattedDate = today.toISOString().split('T')[0];
-        setExpiry(days);
-        setExpiryDate(formattedDate);
-    };
 
     return (
         <MainContainer>
@@ -77,60 +53,8 @@ const SellAndBuyPage = () => {
                             즉시 구매
                         </Button>
                     </ButtonContainer>
-                    {view === 'buy' && (
-                        <BuySection>
-                            <InfoRow>
-                                <InfoLabel>즉시 구매가</InfoLabel>
-                                <InfoValue onClick={() => setView('bid')}>136,000원</InfoValue>
-                            </InfoRow>
-                            <Separator />
-                            <InfoText>총 결제 금액은 다음 화면에서 계산됩니다.</InfoText>
-                            <Separator />
-                            <InfoRow>
-                                <InfoLabel>총 결제금액</InfoLabel>
-                                <Next>다음 화면에서 확인</Next>
-                            </InfoRow>
-                            <Button type='button' buttonColor='dark' size='full'>즉시 구매 계속</Button>
-                        </BuySection>
-                    )}
-                    {view === 'bid' && (
-                        <BidSection>
-                            <InfoRow>
-                                <InfoLabel>구매 희망가</InfoLabel>
-                                <BidInput
-                                    type="text"
-                                    placeholder="희망가 입력"
-                                    value={bidPrice}
-                                    onChange={handleBidPriceChange}
-                                    isError={!!bidError}
-                                />
-                            </InfoRow>
-                            {bidError && <ErrorText>{bidError}</ErrorText>}
-                            <Separator />
-                            <InfoText>총 결제 금액은 다음 화면에서 계산됩니다.</InfoText>
-                            <Separator />
-                            <InfoLabel>입찰 마감기한</InfoLabel>
-                            {expiry !== null && <ExpiryText>{expiry}일 ({expiryDate} 마감)</ExpiryText>}
-                            <ButtonContainer>
-                                {[1, 3, 7, 30, 60, 90, 180].map((days) => (
-                                    <Button
-                                        key={days}
-                                        buttonColor='light'
-                                        size='small'
-                                        onClick={() => handleExpiryChange(days)}
-                                    >
-                                        {days}일
-                                    </Button>
-                                ))}
-                            </ButtonContainer>
-                            <Separator />
-                            <InfoRow>
-                                <InfoLabel>총 결제금액</InfoLabel>
-                                <Next>다음 화면에서 확인</Next>
-                            </InfoRow>
-                            <Button type='button' buttonColor='dark' size='full'>구매 입찰 계속</Button>
-                        </BidSection>
-                    )}
+                    {view === 'buy' && <BuySectionPage />}
+                    {view === 'bid' && <BidSectionPage />}
                 </BottomSection>
             </SellAndBuyContainer>
         </MainContainer>
@@ -234,86 +158,4 @@ const ButtonContainer = styled.div`
     display: flex;
     justify-content: space-between;
     margin-bottom: 1rem;
-`;
-
-const BuySection = styled.div`
-    padding: 1rem;
-    border-radius: 0.5rem;
-    text-align: center;
-`;
-
-const BidSection = styled.div`
-    padding: 1rem;
-    border-radius: 0.5rem;
-    text-align: center;
-`;
-
-const InfoRow = styled.div`
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 0.5rem;
-`;
-
-const InfoLabel = styled.strong`
-    font-size: 1.2rem;
-    color: black;
-    text-align: left;
-`;
-
-const InfoValue = styled.strong`
-    font-size: 1.8rem;
-    margin-bottom: 1rem;
-    cursor: text;
-`;
-
-const InfoText = styled.p`
-    font-size: 1.2rem;
-    color: gray;
-    text-align: left;
-    margin-bottom: 2rem;
-`;
-
-const Next = styled.p`
-    font-size: 1.2rem;
-    color: ${theme.colors.gray[200]};
-`;
-
-interface BidInputProps {
-    isError: boolean;
-}
-
-const BidInput = styled.input<BidInputProps>`
-    font-size: 1.2rem;
-    padding: 0.5rem;
-    border: none;
-    border-bottom: 1px solid ${props => props.isError ? 'red' : 'black'};
-    &:focus {
-        outline: none;
-        border-bottom: 1px solid ${props => props.isError ? 'red' : 'black'};
-    }
-`;
-
-const ErrorText = styled.p`
-    font-size: 1rem;
-    color: red;
-    text-align: right;
-    margin-top: 0;
-    margin-bottom: 1rem;
-`;
-
-const BidButton = styled.button`
-    background: ${theme.colors.gray[200]};
-    border: none;
-    padding: 0.5rem 1rem;
-    margin-right: 0.5rem;
-    cursor: pointer;
-    &:hover {
-        background: ${theme.colors.gray[300]};
-    }
-`;
-
-const ExpiryText = styled.p`
-    font-size: 1.2rem;
-    color: black;
-    margin-top: 0.5rem;
 `;
